+++ totalsegmentator/python_api.py
@@ -71,7 +71,7 @@ def totalsegmentator(input: Union[str, Path, Nifti1Image], output: Union[str, Pa
                      skip_saving=False, device="gpu", license_number=None,
                      statistics_exclude_masks_at_border=True, no_derived_masks=False,
                      v1_order=False, fastest=False, roi_subset_robust=None, stats_aggregation="mean",
-                     remove_small_blobs=False):
+                     remove_small_blobs=False, statistics_normalized_intensities=False):
     """
     Run TotalSegmentator from within python.
 
@@ -626,7 +626,8 @@ def totalsegmentator(input: Union[str, Path, Nifti1Image], output: Union[str, Pa
                             quiet=quiet, verbose=verbose, test=test, skip_saving=skip_saving, device=device,
                             exclude_masks_at_border=statistics_exclude_masks_at_border,
                             no_derived_masks=no_derived_masks, v1_order=v1_order,
-                            stats_aggregation=stats_aggregation, remove_small_blobs=remove_small_blobs)
+                            stats_aggregation=stats_aggregation, remove_small_blobs=remove_small_blobs,
+                            normalized_intensities=statistics_normalized_intensities)
     seg = seg_img.get_fdata().astype(np.uint8)
 
     try:
@@ -650,7 +651,9 @@ def totalsegmentator(input: Union[str, Path, Nifti1Image], output: Union[str, Pa
             stats_file = None
         stats = get_basic_statistics(seg, ct_img, stats_file, 
                                      quiet, task, statistics_exclude_masks_at_border,
-                                     roi_subset, metric=stats_aggregation)
+                                     roi_subset, 
+                                     metric=stats_aggregation,
+                                     normalized_intensities=statistics_normalized_intensities)
         # get_radiomics_features_for_entire_dir(input, output, output / "statistics_radiomics.json")
         if not quiet: print(f"  calculated in {time.time()-st:.2f}s")